test: move requireFresh/cleanCache helper to separate file for re-use
