+++ test/helpers/module-cache-helpers.js
@@ -1,0 +1,13 @@
+function cleanCache (module) {
+  delete require.cache[require.resolve(module)]
+}
+
+function requireFresh (module) {
+  cleanCache(module)
+  return require(module)
+}
+
+module.exports = {
+  cleanCache,
+  requireFresh
+}
